support nested value in updatedsk
